🚚 rename: formatData 함수 분리 #40
