+++ src/pages/Home/utils/formatDate.ts
@@ -1,0 +1,9 @@
+export default function formatDate(dateString: string): string {
+  const date = new Date(dateString);
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  const daysOfWeek = ["일", "월", "화", "수", "목", "금", "토"];
+  const weekday = daysOfWeek[date.getDay()];
+
+  return `${month}/${day}(${weekday})`;
+}
